enchancement: refactoring workflows
